ref(replay): Refactor the replay memory chart and render correct start/end times
