+++ static/app/views/replays/detail/layout/focusArea.tsx
@@ -1,22 +1,15 @@
-import styled from '@emotion/styled';
-
-import {space} from 'sentry/styles/space';
 import useActiveReplayTab, {TabKey} from 'sentry/utils/replays/hooks/useActiveReplayTab';
 import A11y from 'sentry/views/replays/detail/accessibility/index';
 import Breadcrumbs from 'sentry/views/replays/detail/breadcrumbs';
 import Console from 'sentry/views/replays/detail/console';
-import DomNodesChart from 'sentry/views/replays/detail/domNodesChart';
 import ErrorList from 'sentry/views/replays/detail/errorList/index';
-import FluidHeight from 'sentry/views/replays/detail/layout/fluidHeight';
-import MemoryChart from 'sentry/views/replays/detail/memoryChart';
+import MemoryPanel from 'sentry/views/replays/detail/memoryPanel/index';
 import NetworkList from 'sentry/views/replays/detail/network';
 import PerfTable from 'sentry/views/replays/detail/perfTable/index';
 import TagPanel from 'sentry/views/replays/detail/tagPanel';
 import Trace from 'sentry/views/replays/detail/trace/index';
 
-type Props = {};
-
-function FocusArea({}: Props) {
+export default function FocusArea() {
   const {getActiveTab} = useActiveReplayTab();
 
   switch (getActiveTab()) {
@@ -31,12 +24,7 @@ function FocusArea({}: Props) {
     case TabKey.ERRORS:
       return <ErrorList />;
     case TabKey.MEMORY:
-      return (
-        <MemoryTabWrapper>
-          <MemoryChart />
-          <DomNodesChart />
-        </MemoryTabWrapper>
-      );
+      return <MemoryPanel />;
     case TabKey.CONSOLE:
       return <Console />;
     case TabKey.TAGS:
@@ -47,12 +35,3 @@ function FocusArea({}: Props) {
     }
   }
 }
-
-const MemoryTabWrapper = styled(FluidHeight)`
-  justify-content: center;
-  gap: ${space(1)};
-  height: 100%;
-  display: flex;
-`;
-
-export default FocusArea;
